extract RepositoryModel and RepositoryProvider
