+++ packages/patapata_core/lib/src/repository.dart
@@ -6,7 +6,6 @@
 import 'dart:collection';
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:patapata_core/patapata_annotation.dart';
 import 'package:patapata_core/patapata_core.dart';
@@ -15,9 +14,10 @@ import 'package:patapata_core/patapata_interface.dart';
 import 'package:provider/provider.dart';
 
 import 'exception.dart';
-import 'synchronous_future.dart';
 
 part 'repository_cache_map.dart';
+part 'repository_model.dart';
+part 'repository_provider.dart';
 
 final _logger = Logger('patapata.Repository');
 
@@ -38,246 +38,6 @@ enum RepositoryFetchPolicy {
 
   /// Do not use the cache.
   noCache,
-}
-
-/// Base class that serves as the foundation for the repository model.
-/// [RepositoryModel] and [SimpleRepositoryModel] inherit from this class.
-mixin RepositoryModelBase<T extends RepositoryModelBase<T, I>,
-    I extends Object> {
-  /// The unique ID in the repository.
-  I get repositoryId;
-
-  /// A map of variables entrusted to the repository for management.
-  /// Set the keys of the map to the sets specified in [RepositoryClass].
-  /// Please configure the combinations of parameter sets and variables used in each parameter set.
-  ///
-  /// For example:
-  /// ```dart
-  ///
-  /// @override
-  /// Map<Type, Set<ProviderModelVariable>> get repositorySetVariables => {
-  ///       Set1: {_value1, _value2},
-  ///       Set2: {_value1, _value2, _value3},
-  ///     };
-  /// ```
-  Map<Type, Set<ProviderModelVariable>> get repositorySetVariables;
-
-  /// Called by a Repository to create a default object of type [T] from [id].
-  T repositoryDefaultFactory(I id);
-
-  /// Function for creating a widget that encapsulates a data model.
-  /// Add it to the Widget tree using [Provider] or [InheritedWidget], etc.
-  /// You may need to change the created widget depending on the type of data model and how it is used.
-  ///
-  /// For example, if you are using [RepositoryClass.sets], you need to prepare a data model and a [MultiProvider] for the set.
-  /// For example:
-  /// ```dart
-  ///
-  /// @override
-  /// Widget providersBuilder(Widget child) {
-  ///   return MultiProvider(
-  ///     providers: [
-  ///       InheritedProvider<DataModel>.value(
-  ///         value: this,
-  ///         startListening: (c, v) {
-  ///           v.addListener(c.markNeedsNotifyDependents);
-  ///           return () => v.removeListener(c.markNeedsNotifyDependents);
-  ///         },
-  ///       ),
-  ///       InheritedProvider<DataModelSet>.value(
-  ///         value: this,
-  ///         startListening: (c, v) {
-  ///           final tInstance = (v as DataModel);
-  ///           tInstance.addListener(c.markNeedsNotifyDependents);
-  ///           return () => tInstance.removeListener(c.markNeedsNotifyDependents);
-  ///         },
-  ///       ),
-  ///     ],
-  ///     child: child,
-  ///   );
-  /// }
-  /// ```
-  ///
-  /// If you prepare data using global (such as riverpod), there is no need to override it.
-  /// However, since the repository system needs to understand how to retrieve data,
-  /// please specify that method in [RepositoryProvider.reader].
-  Widget providersBuilder(Widget child);
-
-  /// Returns a list of sets specified in [RepositoryClass].
-  Set<Type> get repositorySets;
-
-  /// Checks if the specified parameter set is included in [repositorySets].
-  bool repositoryHasSet(Type set);
-
-  /// Returns whether the repository has change notifications.
-  bool get _repositoryHasListeners;
-
-  /// Notifies the repository of changes.
-  void _repositoryNotifyListeners();
-
-  /// Clones this [RepositoryObjectBase] object.
-  /// Setting deep to true will also copy the contents if the source is [RepositoryModelBase].
-  T clone([bool deep = true]);
-
-  static T2 _clone<T2 extends RepositoryModelBase<T2, I>, I extends Object>(
-    RepositoryModelBase<T2, I> original, [
-    bool deep = true,
-  ]) {
-    final tClone = original.repositoryDefaultFactory(original.repositoryId);
-    if (tClone is! ProviderModel) {
-      if (tClone is SimpleRepositoryModel) {
-        (tClone as SimpleRepositoryModel).update(original);
-      }
-
-      return tClone;
-    }
-
-    final tCloneBatch = (tClone as ProviderModel).begin();
-
-    for (var i in original.repositorySetVariables.entries) {
-      final tCloneSetVariables = tClone.repositorySetVariables[i.key]!;
-      final tOriginalSetVariables = i.value;
-
-      for (var j = 0, jl = tCloneSetVariables.length; j < jl; j++) {
-        final tOriginalVariable = tOriginalSetVariables.elementAt(j);
-
-        if (tOriginalVariable.set) {
-          final tOriginalValue = tOriginalVariable.unsafeValue;
-
-          tCloneBatch.set(
-            tCloneSetVariables.elementAt(j),
-            deep && tOriginalValue is RepositoryModelBase
-                ? tOriginalValue.clone(true)
-                : tOriginalValue,
-          );
-        }
-      }
-    }
-
-    tCloneBatch.commit(notify: false);
-
-    return tClone;
-  }
-}
-
-/// When creating a repository model that requires change notifications, mix-in this class.
-/// It is assumed that the inheriting class is a [ProviderModel].
-mixin RepositoryModel<T extends RepositoryModelBase<T, I>, I extends Object>
-    on ProviderModel<T> implements RepositoryModelBase<T, I> {
-  /// Updates the variables held by the repository.
-  /// The data is copied from [object].
-  /// Using [batch] allows for bulk updates and simultaneous change notifications.
-  void update(ProviderModelBatch batch, T object) {
-    final tVariables = repositorySetVariables[T]!;
-    final tObjectVariables =
-        object.repositorySetVariables[T]!.toList(growable: false);
-
-    for (final v in tObjectVariables.where((e) => e.set)) {
-      final i = tObjectVariables.indexOf(v);
-      batch.set(tVariables.elementAt(i), v.unsafeValue);
-    }
-  }
-
-  @override
-  Set<Type> get repositorySets => {
-        for (var i in repositorySetVariables.entries)
-          if (i.value.every((e) => e.set)) i.key,
-      };
-
-  @override
-  bool repositoryHasSet(Type set) =>
-      repositorySetVariables[set]?.every((e) => e.set) ?? false;
-
-  @override
-  Widget providersBuilder(Widget child) => child;
-
-  @override
-  bool get _repositoryHasListeners => hasListeners;
-
-  @override
-  void _repositoryNotifyListeners() {
-    if (hasListeners) {
-      scheduleFunction(() {
-        if (hasListeners) {
-          notifyListeners();
-        }
-      });
-    }
-  }
-
-  @override
-  T clone([bool deep = true]) {
-    return RepositoryModelBase._clone<T, I>(this, deep);
-  }
-
-  @override
-  int get hashCode => repositoryId.hashCode;
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) {
-      return true;
-    }
-
-    if (other is T && other.repositoryId == repositoryId) {
-      return true;
-    }
-
-    return false;
-  }
-}
-
-/// Mix-in when creating a simple repository model that does not require change notifications.
-/// It is possible to have change notifications, but it is often preferable to use [RepositoryModel].
-/// Use this when, due to the structure of the data model, [RepositoryModel] cannot be used.
-mixin SimpleRepositoryModel<T extends RepositoryModelBase<T, I>,
-    I extends Object> implements RepositoryModelBase<T, I> {
-  /// Updates the variables held by the repository.
-  /// The specifics of the update are left to the derived class.
-  void update(T object);
-
-  @override
-  Set<Type> get repositorySets => {T};
-
-  @override
-  Map<Type, Set<ProviderModelVariable>> get repositorySetVariables => {
-        T: {},
-      };
-
-  // coverage:ignore-start
-  // This method is not used in SimpleRepositoryModel.
-  @override
-  Widget providersBuilder(Widget child) => child;
-  // coverage:ignore-end
-
-  @override
-  bool repositoryHasSet(Type set) => true;
-
-  @override
-  void _repositoryNotifyListeners() {
-    // Do nothing.
-  }
-
-  @override
-  T clone([deep = true]) {
-    return RepositoryModelBase._clone<T, I>(this, deep);
-  }
-
-  @override
-  int get hashCode => repositoryId.hashCode;
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) {
-      return true;
-    }
-
-    if (other is T && other.repositoryId == repositoryId) {
-      return true;
-    }
-
-    return false;
-  }
 }
 
 class _FetchSetEntry<I extends Object> {
@@ -909,466 +669,5 @@ abstract class Repository<T extends RepositoryModelBase<T, I>,
     for (var i in tObjects) {
       _removeCache(i);
     }
-  }
-}
-
-class _SynchronousTracker {
-  bool callerFinished = false;
-  bool calledFinished = false;
-}
-
-typedef _RepositoryDataHolderBuilder<T> = Widget Function(
-    BuildContext context, T? data);
-
-class _RepositoryCore<T extends RepositoryModelBase<T, I>, I extends Object,
-    R extends Repository<T, I>> extends StatefulWidget {
-  final I id;
-  final Future<T?> Function(R repository) fetcher;
-  final R repository;
-  final _RepositoryDataHolderBuilder<T> holderBuilder;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-
-  const _RepositoryCore({
-    super.key,
-    required this.id,
-    required this.fetcher,
-    required this.repository,
-    required this.holderBuilder,
-    this.loading,
-    this.errorBuilder,
-  });
-
-  @override
-  State<_RepositoryCore<T, I, R>> createState() =>
-      _RepositoryCoreState<T, I, R>();
-}
-
-class _RepositoryCoreState<
-    T extends RepositoryModelBase<T, I>,
-    I extends Object,
-    R extends Repository<T, I>> extends State<_RepositoryCore<T, I, R>> {
-  bool? _loading;
-  Object? _error;
-  StackTrace? _stackTrace;
-  T? _data;
-
-  void _updateData(_SynchronousTracker tracker) {
-    _loading = true;
-    _data = null;
-    _error = null;
-    _stackTrace = null;
-
-    runZoned(() {
-      widget.fetcher(widget.repository).then((data) {
-        _data = data;
-
-        return data;
-      }).catchError((e, stackTrace) {
-        _error = e;
-        _stackTrace = stackTrace;
-
-        throw e;
-      }).whenComplete(() {
-        _loading = false;
-        tracker.calledFinished = true;
-
-        if (tracker.callerFinished) {
-          if (mounted) {
-            setState(() {});
-          }
-        }
-      });
-    }, zoneValues: {
-      #repositorySynchronousCache: true,
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final Widget tWidget;
-    final Key tKey;
-
-    if (_loading == null) {
-      final tTracker = _SynchronousTracker();
-      _updateData(tTracker);
-      tTracker.callerFinished = true;
-    }
-
-    if (_loading == true) {
-      tKey = const ValueKey('loading');
-      tWidget = widget.loading ?? const SizedBox.shrink();
-    } else if (_error != null) {
-      tKey = const ValueKey('error');
-      tWidget = widget.errorBuilder?.call(context, _error!, _stackTrace) ??
-          const SizedBox.shrink();
-    } else if (_data == null) {
-      tKey = const ValueKey('error');
-      tWidget = widget.errorBuilder?.call(
-              context, ProviderNullException(null.runtimeType, T), null) ??
-          const SizedBox.shrink();
-    } else {
-      if (_data is Listenable) {
-        tKey = ValueKey('listenable:${_data.hashCode}');
-        tWidget = widget.holderBuilder(context, _data);
-      } else {
-        tKey = ValueKey('nonListenable:${_data.hashCode}');
-        tWidget = widget.holderBuilder(context, _data);
-      }
-    }
-
-    return KeyedSubtree(
-      key: tKey,
-      child: tWidget,
-    );
-  }
-}
-
-class _RepositoryMultiCore<T extends RepositoryModelBase<T, I>,
-    I extends Object, R extends Repository<T, I>> extends StatefulWidget {
-  final Future<List<T?>> Function(R repository) fetcher;
-  final R repository;
-  final _RepositoryDataHolderBuilder<List<T?>> holderBuilder;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-
-  const _RepositoryMultiCore({
-    super.key,
-    required this.fetcher,
-    required this.repository,
-    required this.holderBuilder,
-    this.loading,
-    this.errorBuilder,
-  });
-
-  @override
-  State<_RepositoryMultiCore<T, I, R>> createState() =>
-      _RepositoryMultiCoreState<T, I, R>();
-}
-
-class _RepositoryMultiCoreState<
-    T extends RepositoryModelBase<T, I>,
-    I extends Object,
-    R extends Repository<T, I>> extends State<_RepositoryMultiCore<T, I, R>> {
-  bool? _loading;
-  Object? _error;
-  StackTrace? _stackTrace;
-  List<T?>? _data;
-
-  void _updateData(_SynchronousTracker tracker) {
-    _loading = true;
-    _data = null;
-    _error = null;
-    _stackTrace = null;
-
-    runZoned(() {
-      widget.fetcher(widget.repository).then((data) {
-        _data = data;
-
-        return data;
-      }).catchError((e, stackTrace) {
-        _error = e;
-        _stackTrace = stackTrace;
-
-        throw e;
-      }).whenComplete(() {
-        _loading = false;
-        tracker.calledFinished = true;
-
-        if (tracker.callerFinished) {
-          if (mounted) {
-            setState(() {});
-          }
-        }
-      });
-    }, zoneValues: {
-      #repositorySynchronousCache: true,
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final Widget tWidget;
-    final Key tKey;
-
-    if (_loading == null) {
-      final tTracker = _SynchronousTracker();
-      _updateData(tTracker);
-      tTracker.callerFinished = true;
-    }
-
-    if (_loading == true) {
-      tKey = const ValueKey('loading');
-      tWidget = widget.loading ?? const SizedBox.shrink();
-    } else if (_error != null) {
-      tKey = const ValueKey('error');
-      tWidget = widget.errorBuilder?.call(context, _error!, _stackTrace) ??
-          const SizedBox.shrink();
-    } else {
-      assert(_data != null);
-
-      tKey = ValueKey('list:${_data.hashCode}');
-      tWidget = widget.holderBuilder(context, _data);
-    }
-
-    return KeyedSubtree(
-      key: tKey,
-      child: tWidget,
-    );
-  }
-}
-
-/// Widget that interacts with [RepositoryModel] and [Repository].
-/// When retrieving data from the lower hierarchy where this is placed,
-/// [RepositoryModel] is provided in the type specified by [RepositoryClass].
-///
-/// The actual stored data depends on the ID passed to [Repository.fetch].
-/// The necessary [Repository] is provided from [fetcher], so use it to retrieve data.
-///
-/// Please pass the means to access [Repository] to [reader].
-/// For example, when obtaining it through [BuildContext] and with the library provider,
-/// it will be "() => context.read<Repository>()".
-class RepositoryProvider<T extends RepositoryModelBase<T, I>, I extends Object,
-    R extends Repository<T, I>> extends StatelessWidget {
-  final I id;
-  final Future<T?> Function(R repository) fetcher;
-  final R repository;
-  final WidgetBuilder builder;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-
-  const RepositoryProvider({
-    super.key,
-    required this.id,
-    required this.fetcher,
-    required this.repository,
-    required this.builder,
-    this.loading,
-    this.errorBuilder,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return _RepositoryCore<T, I, R>(
-      id: id,
-      fetcher: fetcher,
-      repository: repository,
-      holderBuilder: (context, data) {
-        return data!.providersBuilder(Builder(builder: builder));
-      },
-      loading: loading,
-      errorBuilder: errorBuilder,
-    );
-  }
-}
-
-/// A version of [RepositoryProvider] that supports multiple items.
-/// For details, please refer to [RepositoryProvider].
-class RepositoryMultiProvider<
-    T extends RepositoryModelBase<T, I>,
-    I extends Object,
-    R extends Repository<T, I>,
-    S extends Object> extends StatelessWidget {
-  final Future<List<T?>> Function(R repository) fetcher;
-  final R repository;
-  final WidgetBuilder builder;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-
-  const RepositoryMultiProvider({
-    super.key,
-    required this.fetcher,
-    required this.repository,
-    required this.builder,
-    this.loading,
-    this.errorBuilder,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return _RepositoryMultiCore(
-      fetcher: fetcher,
-      repository: repository,
-      holderBuilder: (context, data) {
-        return Provider<List<S>>.value(
-          value: data!.map((e) => e as S).toList(),
-          child: Builder(builder: builder),
-        );
-      },
-      loading: loading,
-      errorBuilder: errorBuilder,
-    );
-  }
-}
-
-typedef RepositoryObserverBuilder<T> = Widget Function(
-    BuildContext context, Widget? child, T? data);
-
-/// The basic functionality is the same as [RepositoryProvider].
-/// The difference is that it is used when change notifications are not needed, such as in the case of [SimpleRepositoryModel],
-/// or when dealing with simple data models.
-class RepositoryObserver<T extends RepositoryModelBase<T, I>, I extends Object,
-    R extends Repository<T, I>> extends StatelessWidget {
-  final I id;
-  final Future<T?> Function(R repository) fetcher;
-  final R repository;
-  final RepositoryObserverBuilder<T> builder;
-  final bool notify;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-  final Widget? child;
-
-  const RepositoryObserver({
-    super.key,
-    required this.id,
-    required this.fetcher,
-    required this.repository,
-    required this.builder,
-    required this.notify,
-    this.loading,
-    this.errorBuilder,
-    this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return _RepositoryCore<T, I, R>(
-      id: id,
-      fetcher: fetcher,
-      repository: repository,
-      holderBuilder: (context, data) {
-        if (data != null && notify) {
-          assert(data is Listenable);
-          if (data is Listenable) {
-            return ListenableBuilder(
-              listenable: data as Listenable,
-              child: child,
-              builder: (context, child) {
-                return builder(context, child, data);
-              },
-            );
-          }
-        }
-
-        return builder(context, child, data);
-      },
-      loading: loading,
-      errorBuilder: errorBuilder,
-    );
-  }
-}
-
-/// A version of [RepositoryObserver] that supports multiple items.
-/// For details, please refer to [RepositoryObserver].
-class RepositoryMultiObserver<T extends RepositoryModelBase<T, I>,
-    I extends Object, R extends Repository<T, I>> extends StatelessWidget {
-  final Future<List<T?>> Function(R repository) fetcher;
-  final R repository;
-  final RepositoryObserverBuilder<List<T?>> builder;
-  final bool notify;
-  final Widget? loading;
-  final Widget Function(
-      BuildContext context, Object error, StackTrace? stackTrace)? errorBuilder;
-  final Widget? child;
-
-  const RepositoryMultiObserver({
-    super.key,
-    required this.fetcher,
-    required this.repository,
-    required this.builder,
-    required this.notify,
-    this.loading,
-    this.errorBuilder,
-    this.child,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return _RepositoryMultiCore(
-      fetcher: fetcher,
-      repository: repository,
-      holderBuilder: (context, data) {
-        if (notify && data != null) {
-          final tListenableList = data.whereType<Listenable>();
-
-          assert(tListenableList.isNotEmpty);
-          if (tListenableList.isNotEmpty) {
-            return _ObjectListenable(
-              listenables: tListenableList.toList(),
-              child: child,
-              builder: (context, child) {
-                return builder(context, child, data);
-              },
-            );
-          }
-        }
-
-        return builder(context, child, data);
-      },
-      loading: loading,
-      errorBuilder: errorBuilder,
-    );
-  }
-}
-
-class _ObjectListenable extends StatefulWidget {
-  const _ObjectListenable({
-    required this.listenables,
-    required this.builder,
-    required this.child,
-  });
-
-  final List<Listenable> listenables;
-  final Widget? child;
-  final TransitionBuilder builder;
-
-  @override
-  State<_ObjectListenable> createState() => _ObjectListenableState();
-}
-
-class _ObjectListenableState extends State<_ObjectListenable> {
-  _ObjectNotifier? _notifier;
-
-  @override
-  void dispose() {
-    _notifier?.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    _notifier?.dispose();
-    return ListenableBuilder(
-      listenable: _notifier = _ObjectNotifier(
-        listenables: widget.listenables,
-      ),
-      builder: widget.builder,
-    );
-  }
-}
-
-class _ObjectNotifier extends ChangeNotifier {
-  _ObjectNotifier({
-    required this.listenables,
-  }) {
-    for (final v in listenables) {
-      v.addListener(notifyListeners);
-    }
-  }
-
-  final List<Listenable> listenables;
-
-  @override
-  void dispose() {
-    for (final v in listenables) {
-      v.removeListener(notifyListeners);
-    }
-
-    super.dispose();
   }
 }
